requirements.txt
numpy==1.24.1
selenium==4.20.0
biopython==1.83